fixed rendering issue after polling while sorting
