+++ src/ui/layouts/stackComponents/Stacks/useService.ts
@@ -47,22 +47,6 @@ export const useService = (): ServiceInterface => {
     );
   }, [locationPath]);
 
-  useEffect(() => {
-    const intervalId = setInterval(() => {
-      //assign interval to a variable to clear it.
-      dispatch(
-        stackComponentsActions.getMy({
-          // id: currentWorkspace.id,
-          type: locationPath.split('/')[2],
-          onSuccess: () => setFetching(false),
-          onFailure: () => setFetching(false),
-        }),
-      );
-    }, 5000);
-
-    return () => clearInterval(intervalId);
-  });
-
   const setFetching = (fetching: boolean) => {
     dispatch(stackPagesActions.setFetching({ fetching }));
   };
